services/auth/requirements.txt
coloredlogs==14.0
Flask==1.1.1
flask_CORS==3.0.7
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask_RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
marshmallow-sqlalchemy==0.23.0
marshmallow_enum==1.5.1
psycopg2-binary==2.8.3
pylint==2.4.4
pytest==5.4.3
requests==2.23.0
SQLAlchemy-Utils==0.36.6
